python/requirements.txt
paho-mqtt==1.6.1
numpy==1.26.4
scikit-learn==1.4.2
pandas==2.2.2
joblib==1.4.2
python-dotenv==1.0.1
